+++ code-review.el
@@ -53,6 +53,7 @@
 
 (defcustom code-review-headers-hook
   '(code-review-section-insert-header-title
+    code-review-section-insert-author
     code-review-section-insert-title
     code-review-section-insert-state
     code-review-section-insert-ref