+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Unreleased
 
 - [#169](https://github.com/wandersoncferreira/code-review/pull/169): [gitlab] set labels command
+- [#175](https://github.com/wandersoncferreira/code-review/pull/175): display PR author in the header. Click/RET to follow
 
 # v0.0.6